replace user auth tutorial with google example demo using passport
