+++ auth-user-pwd-google/app.js
@@ -5,14 +5,11 @@ var express = require('express');
 var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
-var passport = require('passport');
 var session = require('express-session');
-
+var passport = require('passport');
 var SQLiteStore = require('connect-sqlite3')(session);
-
 var indexRouter = require('./routes/index');
 var authRouter = require('./routes/auth');
-
 var app = express();
 
 app.locals.pluralize = require('pluralize');
@@ -33,14 +30,13 @@ app.use(session({
   store: new SQLiteStore({ db: 'sessions.db', dir: './var/db' })
 }));
 app.use(passport.authenticate('session'));
-
 app.use('/', indexRouter);
 app.use('/', authRouter);
-
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
   next(createError(404));
 });
+
 
 // error handler
 app.use(function(err, req, res, next) {
