+++ auth-user-pwd-google/package.json
@@ -43,6 +43,7 @@
     "mkdirp": "^1.0.4",
     "morgan": "~1.9.1",
     "passport": "^0.5.2",
+    "passport-google-oidc": "^0.1.0",
     "passport-local": "^1.0.0",
     "pluralize": "^8.0.0",
     "sqlite3": "^5.0.2"
