--- username-password-tutorial/.gitignore
@@ -1,9 +1,0 @@
-.env
-var
-
-# Node.js
-node_modules/
-npm-debug.log*
-
-# Mac OS X
-.DS_Store
