--- username-password-tutorial/views/login.ejs
@@ -1,38 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<title>Express • TodoMVC</title>
-	<link rel="stylesheet" href="/css/base.css">
-	<link rel="stylesheet" href="/css/index.css">
-	<link rel="stylesheet" href="/css/login.css">
-</head>
-
-<body>
-	<section class="prompt">
-		<h3>todos</h3>
-		<h1>Sign in</h1>
-		<form action="/login/password" method="post">
-			<section>
-				<label for="username">Username</label>
-				<input id="username" name="username" type="text" autocomplete="username" required autofocus>
-			</section>
-			<section>
-				<label for="current-password">Password</label>
-				<input id="current-password" name="password" type="password" autocomplete="current-password" required>
-			</section>
-			<button type="submit">Sign in</button>
-		</form>
-		<hr>
-		<p class="help">Don't have an account? <a href="/signup">Sign up</a></p>
-	</section>
-	<footer class="info">
-		<p>Created by <a href="https://www.jaredhanson.me">Jared Hanson</a></p>
-		<p>Part of <a href="https://todomvc.com">TodoMVC</a></p>
-		<p>Authentication powered by <a href="https://www.passportjs.org">Passport</a></p>
-	</footer>
-</body>
-
-</html>
